Automatically backed up by Learn

+++ db/migrate/20190312151507_create_recipes_table.rb
@@ -3,6 +3,6 @@ class CreateRecipesTable < ActiveRecord::Migration
     create_table :recipes do |t|
       t.string = :name
       t.string = :ingredients
-      t.integer = :cook_time
+      t.float = :cook_time
   end
 end
